Removed dead code, minor renames
