+++ azext_iot/monitor/handlers/common_handler.py
