+++ azext_iot/monitor/parsers/__init__.py
@@ -3,8 +3,3 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
-
-from azext_iot.monitor.handlers.common_handler import CommonHandler
-from azext_iot.monitor.handlers.central_handler import CentralHandler
-
-__all__ = ["CommonHandler", "CentralHandler"]
